chore(ew): make new updates and include monthly
