+++ apps/epic-web/src/inngest/inngest.config.ts
@@ -13,7 +13,6 @@ import {syncDiscordRoles} from 'inngest/functions/discord/sync-discord-roles'
 import {slackDailyReporter} from './functions/stripe/slack-daily-reporter'
 import {slackMonthlyReporter} from './functions/stripe/slack-monthly-reporter'
 import {syncConversions} from 'inngest/functions/sync-conversions'
-import {slackWeeklyReporter} from './functions/stripe/slack-weekly-reporter'
 
 export const inngestConfig = {
   client: inngest,
@@ -34,6 +33,5 @@ export const inngestConfig = {
     slackMonthlyReporter,
     syncConversions,
     ...sanityProductFunctions,
-    slackWeeklyReporter,
   ],
 }
